requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
anthropic==0.42.0
requests==2.32.3
WTForms==3.2.1
email-validator==2.2.0
python-dotenv==1.0.1
gunicorn==23.0.0
